Move search bar to it's own file
