+++ gui/src/renderer/components/SplitTunnelingSettings.tsx
@@ -31,8 +31,6 @@ import {
   StyledCellButton,
   StyledCellLabel,
   StyledCellWarningIcon,
-  StyledClearButton,
-  StyledClearIcon,
   StyledContent,
   StyledHeaderTitle,
   StyledHeaderTitleContainer,
@@ -43,9 +41,7 @@ import {
   StyledNoResult,
   StyledNoResultText,
   StyledPageCover,
-  StyledSearchContainer,
-  StyledSearchIcon,
-  StyledSearchInput,
+  StyledSearchBar,
   StyledSpinnerRow,
 } from './SplitTunnelingSettingsStyles';
 import Switch from './Switch';
@@ -176,7 +172,7 @@ function LinuxSplitTunnelingSettings(props: IPlatformSplitTunnelingSettingsProps
         </HeaderSubTitle>
       </SettingsHeader>
 
-      <SearchBar searchTerm={searchTerm} onSearch={setSearchTerm} />
+      <StyledSearchBar searchTerm={searchTerm} onSearch={setSearchTerm} />
       <ApplicationList applications={filteredApplications} rowRenderer={rowRenderer} />
 
       <StyledBrowseButton onClick={launchWithFilePicker}>
@@ -435,7 +431,9 @@ export function WindowsSplitTunnelingSettings(props: IPlatformSplitTunnelingSett
         </HeaderSubTitle>
       </SettingsHeader>
 
-      {splitTunnelingEnabled && <SearchBar searchTerm={searchTerm} onSearch={setSearchTerm} />}
+      {splitTunnelingEnabled && (
+        <StyledSearchBar searchTerm={searchTerm} onSearch={setSearchTerm} />
+      )}
 
       <Accordion expanded={showSplitSection}>
         <Cell.Section sectionTitle={excludedTitle}>
@@ -563,45 +561,6 @@ function WindowsApplicationRow(props: IWindowsApplicationRowProps) {
         />
       )}
     </Cell.CellButton>
-  );
-}
-
-interface ISearchBarProps {
-  searchTerm: string;
-  onSearch: (searchTerm: string) => void;
-}
-
-function SearchBar(props: ISearchBarProps) {
-  const inputRef = useRef() as React.RefObject<HTMLInputElement>;
-
-  const onInput = useCallback(
-    (event: React.FormEvent) => {
-      const element = event.target as HTMLInputElement;
-      props.onSearch(element.value);
-    },
-    [props.onSearch],
-  );
-
-  const onClear = useCallback(() => {
-    props.onSearch('');
-    inputRef.current?.blur();
-  }, [props.onSearch]);
-
-  return (
-    <StyledSearchContainer>
-      <StyledSearchInput
-        ref={inputRef}
-        value={props.searchTerm}
-        onInput={onInput}
-        placeholder={messages.pgettext('split-tunneling-view', 'Filter...')}
-      />
-      <StyledSearchIcon source="icon-filter" width={24} tintColor={colors.white60} />
-      {props.searchTerm.length > 0 && (
-        <StyledClearButton onClick={onClear}>
-          <StyledClearIcon source="icon-close" width={18} tintColor={colors.white40} />
-        </StyledClearButton>
-      )}
-    </StyledSearchContainer>
   );
 }
